+++ src/internal/m365/controller_test.go
@@ -16,8 +16,6 @@ import (
 	"github.com/alcionai/corso/src/internal/common/idname"
 	inMock "github.com/alcionai/corso/src/internal/common/idname/mock"
 	"github.com/alcionai/corso/src/internal/data"
-	dataMock "github.com/alcionai/corso/src/internal/data/mock"
-	"github.com/alcionai/corso/src/internal/m365/graph"
 	"github.com/alcionai/corso/src/internal/m365/mock"
 	"github.com/alcionai/corso/src/internal/m365/resource"
 	exchMock "github.com/alcionai/corso/src/internal/m365/service/exchange/mock"
@@ -368,49 +366,11 @@ func (suite *ControllerIntegrationSuite) SetupSuite() {
 	ctx, flush := tester.NewContext(t)
 	defer flush()
 
-	suite.ctrl = newController(ctx, t, resource.Users, path.ExchangeService)
+	suite.ctrl = newController(ctx, t, path.ExchangeService)
 	suite.user = tconfig.M365UserID(t)
 	suite.secondaryUser = tconfig.SecondaryM365UserID(t)
 
 	tester.LogTimeOfTest(t)
-}
-
-func (suite *ControllerIntegrationSuite) TestRestoreFailsBadService() {
-	t := suite.T()
-
-	ctx, flush := tester.NewContext(t)
-	defer flush()
-
-	var (
-		restoreCfg = testdata.DefaultRestoreConfig("")
-		sel        = selectors.Selector{
-			Service: selectors.ServiceUnknown,
-		}
-	)
-
-	restoreCfg.IncludePermissions = true
-
-	rcc := inject.RestoreConsumerConfig{
-		BackupVersion:     version.Backup,
-		Options:           control.DefaultOptions(),
-		ProtectedResource: sel,
-		RestoreConfig:     restoreCfg,
-		Selector:          sel,
-	}
-
-	deets, err := suite.ctrl.ConsumeRestoreCollections(
-		ctx,
-		rcc,
-		[]data.RestoreCollection{&dataMock.Collection{}},
-		fault.New(true),
-		count.New())
-	assert.Error(t, err, graph.ErrServiceNotEnabled, clues.ToCore(err))
-	assert.Nil(t, deets)
-
-	status := suite.ctrl.Wait()
-	assert.Equal(t, 0, status.Objects)
-	assert.Equal(t, 0, status.Folders)
-	assert.Equal(t, 0, status.Successes)
 }
 
 func (suite *ControllerIntegrationSuite) TestEmptyCollections() {
@@ -512,7 +472,7 @@ func runRestore(
 
 	start := time.Now()
 
-	restoreCtrl := newController(ctx, t, sci.Resource, path.ExchangeService)
+	restoreCtrl := newController(ctx, t, path.ExchangeService)
 	restoreSel := getSelectorWith(t, sci.Service, sci.ResourceOwners, true)
 
 	rcc := inject.RestoreConsumerConfig{
@@ -581,7 +541,7 @@ func runBackupAndCompare(
 		nameToID[ro] = ro
 	}
 
-	backupCtrl := newController(ctx, t, sci.Resource, path.ExchangeService)
+	backupCtrl := newController(ctx, t, path.ExchangeService)
 	backupCtrl.IDNameLookup = inMock.NewCache(idToName, nameToID)
 
 	backupSel := backupSelectorForExpected(t, sci.Service, expectedDests)
@@ -637,7 +597,6 @@ func runRestoreBackupTest(
 
 	cfg := stub.ConfigInfo{
 		Opts:           opts,
-		Resource:       test.resourceCat,
 		Service:        test.service,
 		Tenant:         tenant,
 		ResourceOwners: resourceOwners,
@@ -683,7 +642,6 @@ func runRestoreTestWithVersion(
 
 	cfg := stub.ConfigInfo{
 		Opts:           opts,
-		Resource:       test.resourceCat,
 		Service:        test.service,
 		Tenant:         tenant,
 		ResourceOwners: resourceOwners,
@@ -721,7 +679,6 @@ func runRestoreBackupTestVersions(
 
 	cfg := stub.ConfigInfo{
 		Opts:           opts,
-		Resource:       test.resourceCat,
 		Service:        test.service,
 		Tenant:         tenant,
 		ResourceOwners: resourceOwners,
@@ -765,9 +722,8 @@ func (suite *ControllerIntegrationSuite) TestRestoreAndBackup_core() {
 
 	table := []restoreBackupInfo{
 		{
-			name:        "EmailsWithAttachments",
-			service:     path.ExchangeService,
-			resourceCat: resource.Users,
+			name:    "EmailsWithAttachments",
+			service: path.ExchangeService,
 			collections: []stub.ColInfo{
 				{
 					PathElements: []string{api.MailInbox},
@@ -792,9 +748,8 @@ func (suite *ControllerIntegrationSuite) TestRestoreAndBackup_core() {
 			},
 		},
 		{
-			name:        "MultipleEmailsMultipleFolders",
-			service:     path.ExchangeService,
-			resourceCat: resource.Users,
+			name:    "MultipleEmailsMultipleFolders",
+			service: path.ExchangeService,
 			collections: []stub.ColInfo{
 				{
 					PathElements: []string{api.MailInbox},
@@ -868,9 +823,8 @@ func (suite *ControllerIntegrationSuite) TestRestoreAndBackup_core() {
 			},
 		},
 		{
-			name:        "MultipleContactsSingleFolder",
-			service:     path.ExchangeService,
-			resourceCat: resource.Users,
+			name:    "MultipleContactsSingleFolder",
+			service: path.ExchangeService,
 			collections: []stub.ColInfo{
 				{
 					PathElements: []string{"Contacts"},
@@ -896,9 +850,8 @@ func (suite *ControllerIntegrationSuite) TestRestoreAndBackup_core() {
 			},
 		},
 		{
-			name:        "MultipleContactsMultipleFolders",
-			service:     path.ExchangeService,
-			resourceCat: resource.Users,
+			name:    "MultipleContactsMultipleFolders",
+			service: path.ExchangeService,
 			collections: []stub.ColInfo{
 				{
 					PathElements: []string{"Work"},
@@ -1027,9 +980,8 @@ func (suite *ControllerIntegrationSuite) TestRestoreAndBackup_core() {
 func (suite *ControllerIntegrationSuite) TestMultiFolderBackupDifferentNames() {
 	table := []restoreBackupInfo{
 		{
-			name:        "Contacts",
-			service:     path.ExchangeService,
-			resourceCat: resource.Users,
+			name:    "Contacts",
+			service: path.ExchangeService,
 			collections: []stub.ColInfo{
 				{
 					PathElements: []string{"Work"},
@@ -1133,7 +1085,7 @@ func (suite *ControllerIntegrationSuite) TestMultiFolderBackupDifferentNames() {
 					restoreCfg.Location,
 				)
 
-				restoreCtrl := newController(ctx, t, test.resourceCat, path.ExchangeService)
+				restoreCtrl := newController(ctx, t, path.ExchangeService)
 
 				rcc := inject.RestoreConsumerConfig{
 					BackupVersion:     version.Backup,
@@ -1167,7 +1119,7 @@ func (suite *ControllerIntegrationSuite) TestMultiFolderBackupDifferentNames() {
 
 			// Run a backup and compare its output with what we put in.
 
-			backupCtrl := newController(ctx, t, test.resourceCat, path.ExchangeService)
+			backupCtrl := newController(ctx, t, path.ExchangeService)
 			backupSel := backupSelectorForExpected(t, test.service, expectedDests)
 			t.Log("Selective backup of", backupSel)
 
@@ -1215,9 +1167,8 @@ func (suite *ControllerIntegrationSuite) TestRestoreAndBackup_largeMailAttachmen
 	subjectText := "Test message for restore with large attachment"
 
 	test := restoreBackupInfo{
-		name:        "EmailsWithLargeAttachments",
-		service:     path.ExchangeService,
-		resourceCat: resource.Users,
+		name:    "EmailsWithLargeAttachments",
+		service: path.ExchangeService,
 		collections: []stub.ColInfo{
 			{
 				PathElements: []string{api.MailInbox},
@@ -1318,7 +1269,7 @@ func (suite *ControllerIntegrationSuite) TestBackup_CreatesPrefixCollections() {
 			defer flush()
 
 			var (
-				backupCtrl = newController(ctx, t, test.resourceCat, path.ExchangeService)
+				backupCtrl = newController(ctx, t, test.service)
 				backupSel  = test.selectorFunc(t)
 				errs       = fault.New(true)
 				start      = time.Now()