+++ src/internal/m365/graph/metadata/metadata_test.go
@@ -73,6 +73,11 @@ var (
 			category: path.PagesCategory,
 			expected: assert.Falsef,
 		},
+		{
+			service:  path.GroupsService,
+			category: path.LibrariesCategory,
+			expected: assert.Truef,
+		},
 	}
 )