+++ src/internal/m365/service/groups/backup.go
@@ -10,8 +10,10 @@ import (
 	"github.com/alcionai/corso/src/internal/common/ptr"
 	"github.com/alcionai/corso/src/internal/data"
 	"github.com/alcionai/corso/src/internal/m365/collection/drive"
+	"github.com/alcionai/corso/src/internal/m365/collection/groups"
 	"github.com/alcionai/corso/src/internal/m365/collection/site"
 	"github.com/alcionai/corso/src/internal/m365/graph"
+	odConsts "github.com/alcionai/corso/src/internal/m365/service/onedrive/consts"
 	"github.com/alcionai/corso/src/internal/m365/support"
 	"github.com/alcionai/corso/src/internal/observe"
 	"github.com/alcionai/corso/src/internal/operations/inject"
@@ -40,6 +42,7 @@ func ProduceBackupCollections(
 		categories           = map[path.CategoryType]struct{}{}
 		ssmb                 = prefixmatcher.NewStringSetBuilder()
 		canUsePreviousBackup bool
+		sitesPreviousPaths   = map[string]string{}
 	)
 
 	ctx = clues.Add(
@@ -77,12 +80,39 @@ func ProduceBackupCollections(
 				Selector:          bpc.Selector,
 			}
 
+			bh := drive.NewGroupBackupHandler(
+				bpc.ProtectedResource.ID(),
+				ptr.Val(resp.GetId()),
+				ac.Drives(),
+				scope)
+
+			cp, err := bh.SitePathPrefix(creds.AzureTenantID)
+			if err != nil {
+				return nil, nil, false, clues.Wrap(err, "getting canonical path")
+			}
+
+			sitesPreviousPaths[ptr.Val(resp.GetId())] = cp.String()
+
 			dbcs, canUsePreviousBackup, err = site.CollectLibraries(
 				ctx,
 				sbpc,
-				drive.NewGroupBackupHandler(bpc.ProtectedResource.ID(), ac.Drives(), scope),
+				bh,
 				creds.AzureTenantID,
 				ssmb,
+				su,
+				errs)
+			if err != nil {
+				el.AddRecoverable(ctx, err)
+				continue
+			}
+
+		case path.ChannelMessagesCategory:
+			dbcs, err = groups.CreateCollections(
+				ctx,
+				bpc,
+				groups.NewChannelBackupHandler(bpc.ProtectedResource.ID(), ac.Channels()),
+				creds.AzureTenantID,
+				scope,
 				su,
 				errs)
 			if err != nil {
@@ -113,5 +143,49 @@ func ProduceBackupCollections(
 		collections = append(collections, baseCols...)
 	}
 
+	// Add metadata about sites
+	md, err := getSitesMetadataCollection(
+		creds.AzureTenantID,
+		bpc.ProtectedResource.ID(),
+		sitesPreviousPaths,
+		su)
+	if err != nil {
+		return nil, nil, false, err
+	}
+
+	collections = append(collections, md)
+
 	return collections, ssmb.ToReader(), canUsePreviousBackup, el.Failure()
+}
+
+func getSitesMetadataCollection(
+	tenantID, groupID string,
+	sites map[string]string,
+	su support.StatusUpdater,
+) (data.BackupCollection, error) {
+	// TODO(meain): Switch to using BuildMetadata
+	// https://github.com/alcionai/corso/pull/4184#discussion_r1316139701
+	p, err := path.Builder{}.ToServiceCategoryMetadataPath(
+		tenantID,
+		groupID,
+		path.GroupsService,
+		path.LibrariesCategory,
+		false)
+	if err != nil {
+		return nil, clues.Wrap(err, "making metadata path")
+	}
+
+	p, err = p.Append(false, odConsts.SitesPathDir)
+	if err != nil {
+		return nil, clues.Wrap(err, "appending sites to metadata path")
+	}
+
+	md, err := graph.MakeMetadataCollection(
+		p,
+		[]graph.MetadataCollectionEntry{
+			graph.NewMetadataEntry(graph.PreviousPathFileName, sites),
+		},
+		su)
+
+	return md, err
 }