+++ src/pkg/services/m365/api/users.go
@@ -20,7 +20,8 @@ import (
 
 // Variables
 var (
-	ErrMailBoxSettingsNotFound = clues.New("mailbox settings not found")
+	ErrMailBoxSettingsNotFound     = clues.New("mailbox settings not found")
+	ErrMailBoxSettingsAccessDenied = clues.New("mailbox settings access denied")
 )
 
 // ---------------------------------------------------------------------------
@@ -120,7 +121,7 @@ func (c Users) GetByID(ctx context.Context, identifier string) (models.Userable,
 		err  error
 	)
 
-	resp, err = c.Stable.Client().Users().ByUserId(identifier).Get(ctx, nil)
+	resp, err = c.Stable.Client().Users().ByUserIdString(identifier).Get(ctx, nil)
 
 	if err != nil {
 		return nil, graph.Wrap(ctx, err, "getting user")
@@ -171,6 +172,7 @@ func appendIfErr(errs []error, err error) []error {
 // Info
 // ---------------------------------------------------------------------------
 
+// TODO(pandeyabs): Remove this once the SDK users have migrated to new APIs
 func (c Users) GetInfo(ctx context.Context, userID string) (*UserInfo, error) {
 	var (
 		// Assume all services are enabled
@@ -219,7 +221,7 @@ func (c Users) GetInfo(ctx context.Context, userID string) (*UserInfo, error) {
 		return userInfo, nil
 	}
 
-	mboxSettings, err := c.getMailboxSettings(ctx, userID)
+	mboxSettings, err := c.GetMailboxSettings(ctx, userID)
 	if err != nil {
 		logger.CtxErr(ctx, err).Info("err getting user's mailbox settings")
 
@@ -229,16 +231,16 @@ func (c Users) GetInfo(ctx context.Context, userID string) (*UserInfo, error) {
 
 		mi.ErrGetMailBoxSetting = append(mi.ErrGetMailBoxSetting, clues.New("access denied"))
 	} else {
-		mi = parseMailboxSettings(mboxSettings, mi)
+		mi = ParseMailboxSettings(mboxSettings, mi)
 	}
 
-	err = c.getFirstInboxMessage(ctx, userID, ptr.Val(inbx.GetId()))
+	err = c.GetFirstInboxMessage(ctx, userID, ptr.Val(inbx.GetId()))
 	if err != nil {
 		if !graph.IsErrQuotaExceeded(err) {
 			return nil, clues.Stack(err)
 		}
 
-		userInfo.Mailbox.QuotaExceeded = graph.IsErrQuotaExceeded(err)
+		mi.QuotaExceeded = graph.IsErrQuotaExceeded(err)
 	}
 
 	userInfo.Mailbox = mi
@@ -266,7 +268,7 @@ func EvaluateMailboxError(err error) error {
 	return err
 }
 
-func (c Users) getMailboxSettings(
+func (c Users) GetMailboxSettings(
 	ctx context.Context,
 	userID string,
 ) (models.Userable, error) {
@@ -290,9 +292,9 @@ func (c Users) GetMailInbox(
 	inbox, err := c.Stable.
 		Client().
 		Users().
-		ByUserId(userID).
+		ByUserIdString(userID).
 		MailFolders().
-		ByMailFolderId(MailInbox).
+		ByMailFolderIdString(MailInbox).
 		Get(ctx, nil)
 	if err != nil {
 		return nil, graph.Wrap(ctx, err, "getting MailFolders")
@@ -308,7 +310,7 @@ func (c Users) GetDefaultDrive(
 	d, err := c.Stable.
 		Client().
 		Users().
-		ByUserId(userID).
+		ByUserIdString(userID).
 		Drive().
 		Get(ctx, nil)
 	if err != nil {
@@ -323,7 +325,7 @@ func (c Users) GetDefaultDrive(
 // exceeded error. Ideally(if available) we should convert this to
 // pull the user's usage via an api and compare if they have used
 // up their quota.
-func (c Users) getFirstInboxMessage(
+func (c Users) GetFirstInboxMessage(
 	ctx context.Context,
 	userID, inboxID string,
 ) error {
@@ -337,9 +339,9 @@ func (c Users) getFirstInboxMessage(
 	_, err := c.Stable.
 		Client().
 		Users().
-		ByUserId(userID).
+		ByUserIdString(userID).
 		MailFolders().
-		ByMailFolderId(inboxID).
+		ByMailFolderIdString(inboxID).
 		Messages().
 		Delta().
 		Get(ctx, config)